Arv/presence state (#21)
This adds a function called `generatePresence` that generates a
collection of presence keys at `-/p/${clientID}/name/${id}`. The type of
these entities are called `PresenceEntity` and are defined as:

```ts
type PresenceEntity = {
  clientID: string;
  id: string;
};
```

When looking up a presence state value, if the `id` is omitted, the `id`
is set to `''` and if the `clientID` is omitted, the `clientID` is set
to the `clientID` of the current client.

When writing and mutating the value the `id` and the `clientID` are
added to the value if they are missing.

You can only mutate the presence state value for the current client. If
you try to mutate a value that has different `clientID` than the current
client a runtime error is thrown.