feat: generatePresence take 2 (#25)
Now the keys that we store the presence entities allow distinguishing if there is an id with the value `''` vs no id provided. The keys used are:

```
-/p/${clientID}/${entityName}/
-/p/${clientID}/${entityName}/id/${id}
```

Then we also modify the types to not allow an `id` in the params passed to `get`, `list` `startAtID` etc if the type does not have an `id`.

We never add an `id` to an entity.