add a `messageUpstream` method to allow retrieval of historical data
When a new view is materialized, it won't have any historical data in it and
is only set up to receive new updates.

We need to tell source(s) that a new view is attached and that they should send
all history down to the newly attached view.

`messageUpstream` allows us to do this by providing a path for leaves to communicate up to sources.

In the future, `messageUpstream` will gather hoistable expressions along the path. I.e., expressions
that we can run directly in the source.

A common case would be:

```
select * from foo where id > x
```

the `id > x` comparison should be hoisted up to the source so the source only sends
down rows that match rather than a full table scan.

non-hoistable expressions are things that are not available in a source
or source index. E.g., filtering on any column that is not a primary key today.