Flesh out what we'll hoist to sources
PullMsg / PullHistoricalData asks a source to send old data.

Sources of history may not need to send _all_ history. And we wouldn't
want them to if they have say a million items.

To deal with that, the graph collects information
 about what could constrain the source data. It does this by augmenting the PullMsg.

 A view sets:
 - ordering
 - query type (optionally)

 Upstream operators set:
 - hoistedConditions

Why?

Querying historical data vs responding to changes in
data are slightly different problems.

E.g.,

 "Find me all items in the set greater than Y" /  SELECT * FROM set WHERE item > Y

vs

 "Find me all queries that do not care about the value of Y
 or where Y is less then item"

Pulling answers the former. The data flow graph
answers the latter as well as being able to (no optimally) answer the former.

The former doesn't have to be answered fully since the graph
will filter any over-fetched rows. This means that the source
only needs to select and apply the constraint that matches
its ordering.