fix: Reuse EntityQuery instances
Now we see if there exists a EntityQuery that matches the AST. If so, we
return that.

To ensure we do not leak EntityQuery objects, the EntityQuery is
referenced through a WeakRef in the cache. This means that if the
EntityQuery is garbage collected it will not longer be kept alive by the
lookup cache.

TODO: The lookup is currently keyed by the JSON representation of the
AST which is unnecessarily large. We could compute a numeric hash
instead but native hash implementations on the web platform is all async.

TODO: Normalize the AST so that more objects can be shared.